+++ src/Symfony/Bundle/FrameworkBundle/Command/WorkflowDumpCommand.php
@@ -18,6 +18,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Workflow\Dumper\GraphvizDumper;
+use Symfony\Component\Workflow\Dumper\MermaidDumper;
 use Symfony\Component\Workflow\Dumper\PlantUmlDumper;
 use Symfony\Component\Workflow\Dumper\StateMachineGraphvizDumper;
 use Symfony\Component\Workflow\Marking;
@@ -51,7 +52,7 @@ workflow in different formats
 
 <info>DOT</info>:  %command.full_name% <workflow name> | dot -Tpng > workflow.png
 <info>PUML</info>: %command.full_name% <workflow name> --dump-format=puml | java -jar plantuml.jar -p > workflow.png
-
+<info>MERMAID</info>: %command.full_name% <workflow name> --dump-format=mermaid | mmdc -o workflow.svg
 EOF
             )
         ;
@@ -75,13 +76,20 @@ EOF
             throw new InvalidArgumentException(sprintf('No service found for "workflow.%1$s" nor "state_machine.%1$s".', $serviceId));
         }
 
-        if ('puml' === $input->getOption('dump-format')) {
-            $transitionType = 'workflow' === $type ? PlantUmlDumper::WORKFLOW_TRANSITION : PlantUmlDumper::STATEMACHINE_TRANSITION;
-            $dumper = new PlantUmlDumper($transitionType);
-        } elseif ('workflow' === $type) {
-            $dumper = new GraphvizDumper();
-        } else {
-            $dumper = new StateMachineGraphvizDumper();
+        switch ($input->getOption('dump-format')) {
+            case 'puml':
+                $transitionType = 'workflow' === $type ? PlantUmlDumper::WORKFLOW_TRANSITION : PlantUmlDumper::STATEMACHINE_TRANSITION;
+                $dumper = new PlantUmlDumper($transitionType);
+                break;
+
+            case 'mermaid':
+                $transitionType = 'workflow' === $type ? MermaidDumper::TRANSITION_TYPE_WORKFLOW : MermaidDumper::TRANSITION_TYPE_STATEMACHINE;
+                $dumper = new MermaidDumper($transitionType);
+                break;
+
+            case 'dot':
+            default:
+                $dumper = ('workflow' === $type) ? new GraphvizDumper() : new StateMachineGraphvizDumper();
         }
 
         $marking = new Marking();